Integrate dashboard features into landing page
